moved components in containers directory to the apps folder under the componentes directory
