+++ webapp/js/components/app/NavigationBar.js
