+++ webapp/js/components/display/history/HistoryDisplay.js
@@ -24,8 +24,7 @@ export class HistoryDisplay extends Component {
 
         if (currentParams.key !== prevParams.key && typeof currentParams.key !== 'undefined') {
             getHistoryForKey(currentParams.namespace, currentParams.key);
-        }
-        else if (currentParams.namespace !== prevParams.namespace ||
+        } else if (currentParams.namespace !== prevParams.namespace ||
             (typeof currentParams.key === 'undefined') && typeof prevParams.key !== 'undefined') {
             getHistoryForNamespace(currentParams.namespace);
         }
