+++ webapp/js/main.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import store from './store/Store';
-import AppContainer from './containers/AppContainer';
+import AppContainer from 'components/app/AppContainer';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import '../style/layout/layout.scss';
